backend/requirements/common.txt
django==3.2
djangorestframework
djangorestframework-jwt
django-cors-headers
django-pydenticon
pillow